+++ custom_components/solis_sdm630_sniffer/const.py
@@ -6,7 +6,6 @@ CONF_TIMEOUT = "timeout"
 DEFAULT_TOPIC = "solis/rs485/raw"
 DEFAULT_TIMEOUT = 60
 CONF_GRID_IMPORT_SIGN = "grid_import_sign"
-CONF_CREATE_UTILITY_METERS = "create_utility_meters"
 DEFAULT_GRID_IMPORT_SIGN = "positive"
 
 CONF_UPDATE_INTERVAL = "update_interval"